requirements.txt
pytest
requests
allure-pytest
pydantic
python-dotenv
